refactor: Move MenuDrawer component
